# server/docs/swagger.yaml
definitions:
...
  models.AppleCalendarAuth:
    type: object
...
  models.BufferTimeOptions:
    properties:
      enabled:
        type: boolean
      time:
        type: integer
    type: object
  models.CalendarAccount:
    properties:
      appleCalendarAuth:
        $ref: '#/definitions/models.AppleCalendarAuth'
        type: object
      calendarType:
        type: string
      email:
        description: Email is required for all calendar accounts
        type: string
      enabled:
        type: boolean
      oAuth2CalendarAuth:
        $ref: '#/definitions/models.OAuth2CalendarAuth'
        type: object
      picture:
        type: string
      subCalendars:
        additionalProperties:
          $ref: '#/definitions/models.SubCalendar'
        type: object
    type: object
...
  models.CalendarOptions:
    properties:
      bufferTime:
        $ref: '#/definitions/models.BufferTimeOptions'
        type: object
      workingHours:
        $ref: '#/definitions/models.WorkingHoursOptions'
        type: object
    type: object
...
  models.Location:
    properties:
      city:
        type: string
      country_code:
        type: string
      country_name:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      postal:
        type: string
      state:
        type: string
    type: object
  models.OAuth2CalendarAuth:
    type: object
...
  models.SubCalendar:
    properties:
      enabled:
        type: boolean
      name:
        type: string
    type: object
  models.User:
    properties:
      _id:
        description: Profile info
        type: string
      calendarAccounts:
        additionalProperties:
          $ref: '#/definitions/models.CalendarAccount'
        description: |-
          CalendarAccounts is a mapping from {`email_CALENDARTYPE` => CalendarAccount} that contains all the
          additional accounts the user wants to see google calendar events for
        type: object
      calendarOptions:
        $ref: '#/definitions/models.CalendarOptions'
        description: Calendar options
        type: object
      email:
        type: string
      firstName:
        type: string
      hasCustomName:
        description: Whether the user has set a custom name for themselves, i.e. don't
          change their name when they sign in
        type: boolean
      isPremium:
        type: boolean
      lastName:
        type: string
      numEventsCreated:
        type: integer
      picture:
        type: string
      primaryAccountKey:
        description: The calendarAccountKey of the account the user first signed in
          with
        type: string
      stripeCustomerId:
        description: Stripe customer ID
        type: string
      timezoneOffset:
        type: integer
    type: object
  models.WorkingHoursOptions:
    properties:
      enabled:
        type: boolean
      endTime:
        type: number
      startTime:
        type: number
    type: object
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Schej.it!
  license: {}
  title: Schej.it API
  version: "1.0"
paths:
  /analytics/downgrade-user:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the user email
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              email:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Downgrades the specified user to Schej Free
      tags:
      - analytics
  /analytics/monthly-active-event-creators:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start date (YYYY-MM-DD) for the range
        in: query
        name: startDate
        required: true
        type: string
      - description: End date (YYYY-MM-DD) for the range
        in: query
        name: endDate
        required: true
        type: string
      - description: Client's timezone offset in minutes from UTC (e.g., -420 for
          UTC-7)
        in: query
        name: timezoneOffset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              allOf:
              - type: object
              - properties:
                  count:
                    type: integer
                  date:
                    type: string
                type: object
            type: array
        "400":
          description: Invalid date format, range, or timezone offset
          schema:
            allOf:
            - type: object
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - type: object
            - properties:
                error:
                  type: string
              type: object
      summary: Gets the daily count of monthly active event creators over a date range
      tags:
      - analytics
  /analytics/monthly-active-event-creators-with-more-than-x-events:
    get:
      consumes:
      - application/json
      parameters:
      - description: Start date (YYYY-MM-DD) for the range
        in: query
        name: startDate
        required: true
        type: string
      - description: End date (YYYY-MM-DD) for the range
        in: query
        name: endDate
        required: true
        type: string
      - description: Client's timezone offset in minutes from UTC (e.g., -420 for
          UTC-7)
        in: query
        name: timezoneOffset
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              allOf:
              - type: object
              - properties:
                  count:
                    type: integer
                  date:
                    type: string
                type: object
            type: array
        "400":
          description: Invalid date format, range, or timezone offset
          schema:
            allOf:
            - type: object
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Internal server error
          schema:
            allOf:
            - type: object
            - properties:
                error:
                  type: string
              type: object
      summary: Gets the daily count of monthly active event creators over a date range
        with more than x events
      tags:
      - analytics
  /analytics/scanned-poster:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the location that poster was scanned from and
          the url that was scanned
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              location:
                $ref: '#/definitions/models.Location'
              url:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Notifies us when poster QR code has been scanned
      tags:
      - analytics
  /analytics/upgrade-dialog-viewed:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the user id
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              userId:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Notifies us when user has viewed the upgrade dialog
      tags:
      - analytics
  /analytics/upgrade-user:
    post:
      consumes:
      - application/json
      parameters:
      - description: Object containing the user email
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              email:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
      summary: Upgrades the specified user to Schej Premium
      tags:
      - analytics
  /analytics/user/{email}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User email
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Gets the user by email
      tags:
      - analytics
...
swagger: "2.0"